requirements.txt
numpy
pandas
scikit-learn
streamlit==1.15.0
matplotlib==3.5.2
sklearn.cluster